// frontend/src/app/core/generated/circabc/model/pagedAutoUploadConfiguration.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AutoUploadConfiguration } from './autoUploadConfiguration';


export interface PagedAutoUploadConfiguration { 
    data: Array<AutoUploadConfiguration>;
    /**
     * return the total amount of nodes, without paging so the UI can compute the number of pages 
     */
    total: number;
}